Add files via upload
Hi Vance,
I have uploaded the file again, and rechecked the corrections indicated by you previously. In doing so, I realized that it had been my intention to have deleted line 830 which referred to the Permian
            ischart:Paleozoic,
        ischart:Neoproterozoic ;
Permian appeared below the above. It had a comma after it, and the preceding line for the Neoproterozoic had a ; 
In now deleting Permian (as Paleozoic is inserted), I have reinserted the ; after Neoproterozoic
Also, I deleted the ## line re the question of inserting a note.
I have committed directly to the Master.