Add files via upload
1.	How do we deal with hyphens in cases such as with the Clarence-Moreton Basin? dct:hasPart data:CLARENCE_MORETON BASIN , ???

Also:
dct:hasPart data:PALMER_BARRON_SUBPROVINCE , (for the Palmer-Barron Subprovince)

and

dct:isPartOf data:CONNORS_AUBURN_PROVINCE , (for Connors-Auburn)

2.	When beginning a line with “””, does the line need to end with the same “””? I thought that it was necessary, but Nick did not do this in this file, but there were only one or two cases of same. Not sure if this was an oversight or not. I know that Nick had not finished working on the file.

3.	dct:alternative EASTERN_FOLD_BELT , EASTERN_SUCCESSION , <<< In cases such as this, do we need to repeat ‘dct:alternative’ for the EASTERN_SUCCESSION ,

Also here: We are not looking at an alternative term, but a non-preferred term. Is there other dct code that would be better suited? I couldn’t find one.

4.	Ditto point 3 for:
dct:hasPart data:EROMANGA_BASIN , SURAT_BASIN , etc.

5.	Is it possible to give the following, for example, as a time range, in this case Llandovery to Frasnain? Here, we would have maximum and minimum ages given.

     time:hasTime ischart:Frasnian , ischart:MiddleDevonian , ischart:LowerDevonian , ischart:Pridoli , ischart:Ludlow ,  ischart:Wenlock , ischart:Llandovery ;

This would be instead of listing all of the above. This is more of an issue for igneous provinces where magmatic activity was episodic and where the full range of associated rocks have not been isotopically dated.

6.	For the following, note that the Great Australian Superbasin (GASB) contains the listed basins, which fall under RHAETIAN_JURASSIC_CRETACEOUS_BASIN_TYPES

data:GreatAustralianSuperbasin dct:hasPart data:EROMANGA_BASIN , SURAT_BASIN , MULGILDIE_BASIN , CARPENTARIA_BASIN , CLARENCE_MORETON BASIN , NAMBOUR_BASIN , LAURA_BASIN , dct:isPartOf data:RHAETIAN_JURASSIC_CRETACEOUS_BASIN_TYPES , 

And there are basins listed under RHAETIAN_JURASSIC_CRETACEOUS_BASIN_TYPES , 
But not all basins listed under the latter are embraced by the GASB (such as the Stanwell, Styx, Baga). Is this an issue?

Note that I have introduced a class of ‘Basin Types’ (I could not find a better term), as not all basins are simply basins. So, under Basin Types, there are basins, troughs, depressions, grabens, half-grabens, etc. The Features Type vocab would need to be amended, if this proposal is acceptable.

7.	data:AdavaleBasin
        a geof:Basin ,
        geo:Feature ,
        sdo:Landform ;
I note that Nick did not have a comma at the end of the first line above, but, as I have added to such lines, I have placed a comma at the end of the line.

8.	time:hasTime ischart:Silurian,
        ischart:Ordovician,
        ischart:Cambrian,
        ischart:Neoproterozoic ;
For the above, I notice that Nick did not put spaces between the time-rock unit (e.g., Ordovician) and the following comma in the main body of the file,

but, down the bottom of the file, where he had basins with non-spatial info listed, he consistently had, for each, a space between the time-rock unit (e.g., Paleogene) and the comma: e.g.:
data:WoorabindaBasin 
    a geof:Basin , geo:Feature , sdo:Landform ;
    time:hasTime ischart:Paleogene , ischart:Neogene ;    
    sdo:name "Woorabinda Basin"@en ;

Why the difference. Does it matter or not? For the moment, I have left this ‘as is’.

9.	Do we need to introduce MERLIN KEY CODE numbers or the equivalent into this file? And also link where we can to GA’s Provinces? Nick is working for GA on this anyway and will be looking at all linkages from their end.

10.	data:BagaBasin dct:isPartOf data:data:RHAETIAN_JURASSIC_CRETACEOUS_BASIN_TYPES , dct:alternative JIM_CROW_BASIN , dct:description """The basin contains the Baga beds (Geoscience-Australia Stratigraphic Number 81555—Reserved Status, Australian Stratigraphic Units Database), renamed from the Jim Crow beds. The Jim Crow Basin, associated with the latter unit, is proposed to be renamed the Baga Basin (Paul Blake, pers. comm., 29.09.2020). ,

Note that I have a full stop and a comma at the end of the line??!!