Add files via upload
Hi Vance, Kate, David, 
Paul has reviewed the file and found it, at least the geological part of it, in good order, with the suggested addition of a Frasnian age to two features. I have followed through on this, and, thus, the file is available for your review of the syntax. There will need to be a couple of modifications to the features type vocab, to embrace some of my additions to the file, such as the basin-types groupings. If there is a better way of approaching these groupings, please advise. These basin groups will need to appear as such in GeoResGlobe, as will other groupings, but all of this may need some tweaking.