Add files via upload
I have fixed-up the bad-character problem. Initially, I thought that it only involved hyphens (which it mainly did), but it also encompassed quotation marks and, in one instance, a series of three full stops ( … ), to indicate that words were left out in a direct quotation. The strange thing about all of this is that this bad character did not occur in all instances of hyphens, etc., only a few.

I down-loaded the ‘highlight bad chars extension’, but it took time to figure out that it was only for VS Code, not for VS (not showing up for the latter as an available extension). Also, the bad character did not show up in VS, with which I prefer to work (as VS puts all of the spatial in a single line, making work with the file much easier for what I was doing with it). After correcting in VS Code, I went back to VS, to make suggested edits (re scopeNotes), and then rechecked in VS Code, for the bad character. Hopefully, there are now all gone.

I have looked at my ‘verbosity’ in scopeNotes. I have reduced some Notes in length, but have otherwise left them as is. I feel that there is insufficient material for a stand-alone, linked document. David C also said that he was happy not to have the info in different places. But, in agreement, I do see the need to follow advice for a linked document for the biostrat vocabs (plural). There is much info and complexity here, and the sands continually shift (as new info comes in).